+++ packages/button/src/lib/IconButton.test.tsx
@@ -1,0 +1,45 @@
+import { fireEvent, waitFor, render } from '@testing-library/react';
+import { HomeIcon } from '@availity/mui-icon';
+import { IconButton } from './IconButton';
+
+describe('Button', () => {
+  test('should render successfully', () => {
+    const { getByRole } = render(<IconButton title="test"><HomeIcon/></IconButton>);
+    expect(getByRole('button')).toBeTruthy();
+  });
+
+  test('should have accessible name', () => {
+    const { getByRole } = render(<IconButton title="test"><HomeIcon/></IconButton>);
+    const button = getByRole('button');
+
+    expect(button).toHaveAccessibleName('test');
+  });
+
+  test('should render tooltip on hover', async () => {
+    const { getByRole } = render(<IconButton title="test"><HomeIcon/></IconButton>);
+    const button = getByRole('button');
+
+    fireEvent.mouseOver(button);
+
+    await waitFor(() => getByRole('tooltip'))
+    expect(getByRole('tooltip')).toBeTruthy();
+  });
+
+  test('should render tooltip on focus', async () => {
+    const { getByRole } = render(<IconButton title="test"><HomeIcon/></IconButton>);
+    const button = getByRole('button');
+
+    fireEvent.focus(button);
+
+    await waitFor(() => getByRole('tooltip'));
+    expect(getByRole('tooltip')).toBeTruthy();
+  });
+
+  test('should render contained button when filled', () => {
+    const { getByRole } = render(<IconButton title="test" filled><HomeIcon/></IconButton>);
+    const button = getByRole('button');
+
+    expect(button).toBeTruthy();
+    expect(button).toHaveClass('MuiButton-contained');
+  });
+});