+++ packages/button/src/lib/LoadingButton.stories.tsx
@@ -1,12 +1,10 @@
 import { StoryObj, Meta } from '@storybook/react';
+import Grid from '@mui/material/Grid';
+import { SearchIcon } from '@availity/mui-icon';
 import { LoadingButton, LoadingButtonProps } from './LoadingButton'
 
 
 const meta: Meta<typeof LoadingButton> = {
-    /* 👇 The title prop is optional.
-     * See https://storybook.js.org/docs/7.0/react/configure/overview#configure-story-loading
-     * to learn how to generate automatic titles
-     */
     title: 'Components/Button/LoadingButton',
     component: LoadingButton,
     tags: ['autodocs'],
@@ -22,6 +20,16 @@ const meta: Meta<typeof LoadingButton> = {
 
   export const _LoadingButton: Story = {
     render: ({...args}: LoadingButtonProps ) => (
-      <LoadingButton loadingPosition='start' {...args}>Primary</LoadingButton>
+      <Grid container spacing={1}>
+        <Grid item xs={12}>
+          <LoadingButton {...args}>No Icon</LoadingButton>
+        </Grid>
+        <Grid item xs={12}>
+          <LoadingButton {...args} startIcon={<SearchIcon/>} loadingPosition='start'>Start Icon</LoadingButton>
+        </Grid>
+        <Grid item xs={12}>
+          <LoadingButton {...args} endIcon={<SearchIcon/>}>End Icon</LoadingButton>
+        </Grid>
+      </Grid>
     )
   };