+++ packages/theme/src/lib/light-theme.ts
@@ -247,22 +247,22 @@ export const lightTheme = {
       styleOverrides: {
         action: {
           margin: '-3px -3px -3px auto',
-          padding: '0 0 0 11px'
+          padding: '0 0 0 11px',
         },
         icon: {
           opacity: 1,
           padding: '4px',
           margin: 0,
-          fontSize: tokens.fontSizesBody1
+          fontSize: tokens.fontSizesBody1,
         },
         message: {
           color: tokens.colorTextPrimary,
-          padding: 0
+          padding: 0,
         },
         root: {
           fontSize: tokens.fontSizesBody1,
           lineHeight: tokens.lineHeightsBody1,
-          padding: '8px'
+          padding: '8px',
         },
       },
     },
@@ -270,11 +270,52 @@ export const lightTheme = {
       styleOverrides: {
         root: {
           margin: 0,
-          fontWeight: tokens.fontWeightsBold
-        }
-      }
+          fontWeight: tokens.fontWeightsBold,
+        },
+      },
     },
     MuiButton: {
+      defaultProps: {
+        // The props to change the default for.
+        disableRipple: true, // No more ripple, on the whole application 💣!
+        disableElevation: true,
+      },
+      styleOverrides: {
+        root: ({ ownerState } : any) => ({
+          ...(ownerState.variant === 'contained' ?
+            {
+              backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Main` as keyof typeof tokens],
+              color: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Contrast` as keyof typeof tokens],
+              "&:hover": {
+                backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Dark` as keyof typeof tokens],
+                boxShadow: 'none',
+              },
+              "&:focus": {
+                backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Dark` as keyof typeof tokens],
+                outline: '2px solid white',
+                boxShadow: `0 0 0px 4px ${tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Dark` as keyof typeof tokens]}`,
+              },
+              "&:active": {
+                backgroundColor: tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Main` as keyof typeof tokens],
+              }
+            }
+            : {
+              "&:hover": {
+                boxShadow: 'none',
+              },
+              "&:focus": {
+                outline: '2px solid white',
+                boxShadow: `0 0 0px 4px ${tokens[`color${ownerState.color.charAt(0).toUpperCase() + ownerState.color.slice(1)}Main` as keyof typeof tokens]}`,
+              },
+            }
+          ),
+          boxShadow: 'none',
+          fontWeight: tokens.fontWeightsBold,
+          textTransform: 'none'
+        }),
+      },
+    },
+    MuiIconButton: {
       defaultProps: {
         // The props to change the default for.
         disableRipple: true, // No more ripple, on the whole application 💣!
@@ -282,91 +323,43 @@ export const lightTheme = {
       },
       styleOverrides: {
         root: {
-          minWidth: '128px',
-          textTransform: "none"
-        }
-      },
-      variants: [
-        {
-          props: { color: "tertiary", variant: "contained" },
-          style: {
-            backgroundColor: tokens.colorTertiaryMain,
-            color: tokens.colorTertiaryContrast,
-            "&:hover": {
-              backgroundColor: tokens.colorTertiaryDark
-            },
-            "&:focus": {
-              backgroundColor: tokens.colorTertiaryDark,
-              outline: "2px solid white",
-              boxShadow: "0 0 0px 4px rgba(0,0,0,1)"
-            },
-            "&:active": {
-              backgroundColor: tokens.colorTertiaryMain
-            }
+          color: tokens.colorTextPrimary,
+          "&:hover": {
+            backgroundColor: tokens.colorTertiaryDark,
+          },
+          "&:focus": {
+            backgroundColor: tokens.colorTertiaryDark,
+            outline: `2px solid ${tokens.colorCommonWhite}`,
+            boxShadow: `0 0 0px 4px ${tokens.colorSecondaryLight}`,
           }
         },
-        {
-          props: { color: "secondary" },
-            style: {
-            backgroundColor: tokens.colorSecondaryMain,
-            color: tokens.colorSecondaryContrast,
-            "&:hover": {
-              backgroundColor: tokens.colorSecondaryDark
-            },
-            "&:focus": {
-              backgroundColor: tokens.colorSecondaryDarker,
-              outline: "2px solid white",
-              boxShadow: "0 0 0px 4px rgba(0,0,0,1)"
-            },
-            "&:active": {
-              backgroundColor: tokens.colorSecondaryMain
-            }
-          }
-        },
-        {
-          props: { color: "primary" },
-            style: {
-            backgroundColor: tokens.colorPrimaryMain,
-            color: tokens.colorPrimaryContrast,
-            "&:hover": {
-              backgroundColor: tokens.colorPrimaryDark
-            },
-            "&:focus": {
-              backgroundColor: tokens.colorPrimaryMain,
-              outline: "2px solid white",
-              boxShadow: "0 0 0px 4px rgba(21,82,188,1)"
-            },
-            "&:active": {
-              backgroundColor: tokens.colorPrimaryLight
-            }
-          }
-        },
-        {
-          props: { color: "error", variant: "contained" },
-          style: {
-            backgroundColor: tokens.colorErrorMain,
-            color: tokens.colorErrorContrast,
-            "&:hover": {
-              backgroundColor: tokens.colorErrorDark
-            },
-            "&:focus": {
-              backgroundColor: tokens.colorErrorMain,
-              outline: "2px solid white",
-              boxShadow: "0 0 0px 4px rgba(21,82,188,1)"
-            },
-            "&:active": {
-              backgroundColor: tokens.colorErrorLight
-            }
-          }
-        }
-      ]
+      }
     },
-    MuiLoadingButton: {
+    MuiLink: {
+      defaultProps: {
+        underline: 'hover',
+      },
       styleOverrides: {
         root: {
-          position: 'relative'
-        }
-      }
+          fontWeight: tokens.fontWeightsBold,
+          '&:active': {
+            color: tokens.colorTextPrimary,
+          },
+          '&:visited': {
+            color: tokens.colorIndigo600,
+          },
+        },
+      },
+    },
+    MuiPaper: {
+      defaultProps: {
+        variant: 'outlined',
+      },
+      styleOverrides: {
+        root: {
+          borderColor: tokens.colorBackgroundBorder,
+        },
+      },
     },
     MuiSvgIcon: {
       styleOverrides: {
@@ -374,7 +367,16 @@ export const lightTheme = {
           fontSize: 'inherit',
         },
       },
+    },
+    MuiTooltip: {
+      styleOverrides: {
+        tooltip: {
+          backgroundColor: tokens.colorGrey400,
+          fontSize: '.785rem',
+          fontStyle: 'normal',
+          padding: '2px 8px'
+        },
+      },
     }
-  }
-}
-
+  },
+};