+++ packages/badge/CHANGELOG.md
@@ -2,6 +2,19 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.6](https://github.com/Availity/element/compare/@availity/mui-badge@0.1.5...@availity/mui-badge@0.1.6) (2023-06-01)
+
+
+### Bug Fixes
+
+* **mui-badge:** add success color type ([e0de967](https://github.com/Availity/element/commit/e0de967626e2d235701bcc07b3bd8b5e2957c8d2))
+
+## [0.1.5](https://github.com/Availity/element/compare/@availity/mui-badge@0.1.4...@availity/mui-badge@0.1.5) (2023-05-23)
+
+## [0.1.4](https://github.com/Availity/element/compare/@availity/mui-badge@0.1.3...@availity/mui-badge@0.1.4) (2023-05-19)
+
+## [0.1.3](https://github.com/Availity/element/compare/@availity/mui-badge@0.1.2...@availity/mui-badge@0.1.3) (2023-05-19)
+
 ## [0.1.2](https://github.com/Availity/element/compare/@availity/mui-badge@0.1.1...@availity/mui-badge@0.1.2) (2023-03-23)
 
 ## [0.1.1](https://github.com/Availity/element/compare/@availity/mui-badge@0.1.0...@availity/mui-badge@0.1.1) (2023-03-06)