+++ packages/icon/CHANGELOG.md
@@ -2,6 +2,18 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.3.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.2.0...@availity/mui-icon@0.3.0) (2023-06-01)
+
+
+### Features
+
+* **mui-button:** design specs ([6b67b24](https://github.com/Availity/element/commit/6b67b24cdfef68e14daaeba18a5fd7d90af46a09))
+
+
+### Bug Fixes
+
+* **mui-icon:** remove default title ([ab6cce4](https://github.com/Availity/element/commit/ab6cce46037b43daae7f69739737cfb3b38ad8f6))
+
 ## [0.2.0](https://github.com/Availity/element/compare/@availity/mui-icon@0.1.1...@availity/mui-icon@0.2.0) (2023-04-07)